requirements.txt
openai==0.10.2
pyperclip
python-dotenv